pyproject.toml
[tool.poetry]
name = "come-again-question-mark"
version = "0.1.7"
description = "Transcription tool for video or audio to text"
authors = ["Henry Bergström <henrybergstrom@protonmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
vosk = "^0.3.30"
progressbar2 = "^3.53.3"
srt = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "come_again_question_mark/__init__.py:__version__",
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "pip install poetry && poetry build"
